+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/batch/BatchBankStatement.java
@@ -35,7 +35,6 @@ import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
-import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.util.Collection;
 import java.util.List;