+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/PurchaseOrderLineController.java
@@ -18,12 +18,9 @@
 package com.axelor.apps.supplychain.web;
 
 import com.axelor.apps.account.service.app.AppAccountService;
-import com.axelor.apps.purchase.db.PurchaseOrder;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
 import com.axelor.apps.supplychain.service.PurchaseOrderLineServiceSupplychainImpl;
 import com.axelor.exception.AxelorException;
-import com.axelor.exception.db.repo.TraceBackRepository;
-import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
@@ -38,11 +35,7 @@ public class PurchaseOrderLineController {
   public void computeAnalyticDistribution(ActionRequest request, ActionResponse response)
       throws AxelorException {
     PurchaseOrderLine purchaseOrderLine = request.getContext().asType(PurchaseOrderLine.class);
-    PurchaseOrder purchaseOrder = purchaseOrderLine.getPurchaseOrder();
-    if (purchaseOrder == null) {
-      purchaseOrder = request.getContext().getParent().asType(PurchaseOrder.class);
-      purchaseOrderLine.setPurchaseOrder(purchaseOrder);
-    }
+
     if (Beans.get(AppAccountService.class).getAppAccount().getManageAnalyticAccounting()) {
       purchaseOrderLine =
           purchaseOrderLineServiceSupplychainImpl.computeAnalyticDistribution(purchaseOrderLine);
@@ -53,19 +46,10 @@ public class PurchaseOrderLineController {
   public void createAnalyticDistributionWithTemplate(ActionRequest request, ActionResponse response)
       throws AxelorException {
     PurchaseOrderLine purchaseOrderLine = request.getContext().asType(PurchaseOrderLine.class);
-    PurchaseOrder purchaseOrder = purchaseOrderLine.getPurchaseOrder();
-    if (purchaseOrder == null) {
-      purchaseOrder = request.getContext().getParent().asType(PurchaseOrder.class);
-      purchaseOrderLine.setPurchaseOrder(purchaseOrder);
-    }
-    if (purchaseOrderLine.getAnalyticDistributionTemplate() != null) {
-      purchaseOrderLine =
-          purchaseOrderLineServiceSupplychainImpl.createAnalyticDistributionWithTemplate(
-              purchaseOrderLine);
-      response.setValue("analyticMoveLineList", purchaseOrderLine.getAnalyticMoveLineList());
-    } else {
-      throw new AxelorException(
-          TraceBackRepository.CATEGORY_CONFIGURATION_ERROR, I18n.get("No template selected"));
-    }
+
+    purchaseOrderLine =
+        purchaseOrderLineServiceSupplychainImpl.createAnalyticDistributionWithTemplate(
+            purchaseOrderLine);
+    response.setValue("analyticMoveLineList", purchaseOrderLine.getAnalyticMoveLineList());
   }
 }